+++ src/transfer/transfer.ts
@@ -1,13 +1,12 @@
-import { readFile } from "fs-extra";
 import { Uri } from "vscode";
-import { saveFileService } from "../miiservice/savefileservice";
 import { SystemConfig, UserConfig, configManager } from "../modules/config";
-import { GetAllFilesInDir, GetRemotePath, ValidatePath } from "../modules/file";
 import { ShowQuickPick } from "../modules/vscode";
 import logger from "../ui/logger";
 import { QuickPickItem } from "../ui/quickpick";
 import statusBar, { Icon } from "../ui/statusbar";
 import { GetUserManager } from "../user/usermanager";
+import { UploadFolderLimited } from "./upload";
+import path = require("path");
 
 
 export async function TransferFolder(uri: Uri, userConfig: UserConfig) {
@@ -25,23 +24,10 @@ export async function TransferFolder(uri: Uri, userConfig: UserConfig) {
         if(!await user.login()) return;
 
         statusBar.updateBar('Transfering', Icon.spinLoading, { duration: -1 });
-        logger.info("Transfer Folder Started");        
+        logger.infos("Transfer Folder", path.basename(uri.fsPath) +": Started"); 
 
-        const localFiles = await GetAllFilesInDir(uri.fsPath);
-        const promises: Promise<any>[] = [];
-        for (let index = 0; index < localFiles.length; index++) {
-            const localFile = localFiles[index];
-            if (await ValidatePath(localFile, userConfig)) {
-                promises.push(readFile(localFile).then((content)=>{
-                    const sourcePath = GetRemotePath(localFile, userConfig);
-                    const base64Content = encodeURIComponent(content.length != 0 ? content.toString('base64') : Buffer.from(" ").toString('base64'));
-                    return saveFileService.call({ host: system.host, port: system.port, body: "Content=" + base64Content }, sourcePath);
-                }));
-            }
-        }
-        
-        await Promise.all(promises);
-        statusBar.updateBar('Done', Icon.success, { duration: 1 });
-        logger.info("Transfer Folder Completed");
+        await UploadFolderLimited(uri.fsPath, userConfig, system);
+        statusBar.updateBar('Transferred', Icon.success, { duration: 1 });
+        logger.infos("Transfer Folder", path.basename(uri.fsPath) +": Completed"); 
     }
 }